Update packages from io.github.simonschiller.permissioncheck to com.telefonica.manifestcheck (#11)
* Add support for AGP 9.0.0

* Update github actions

* Update version number

* Update KDoc

* Update buildSrc/src/main/kotlin/Dependencies.kt


* Update junit-platform-launcher version (copilot suggestion)

* Remove TaskConfiguratorFactory

* Update packages from io.github.simonschiller.permissioncheck to com.telefonica.manifestcheck

* Update packages from io.github.simonschiller.permissioncheck to com.telefonica.manifestcheck

* Update packages from io.github.simonschiller.permissioncheck to com.telefonica.manifestcheck

---------
