+++ plugin-core/src/main/kotlin/com/telefonica/manifestcheck/internal/BaselineHandler.kt
@@ -1,0 +1,106 @@
+package com.telefonica.manifestcheck.internal
+
+import com.telefonica.manifestcheck.internal.data.BasePermission
+import com.telefonica.manifestcheck.internal.data.Feature
+import com.telefonica.manifestcheck.internal.data.Permission
+import com.telefonica.manifestcheck.internal.data.Sdk23Permission
+import com.telefonica.manifestcheck.internal.util.appendElement
+import com.telefonica.manifestcheck.internal.util.createDocumentBuilder
+import com.telefonica.manifestcheck.internal.util.forEach
+import com.telefonica.manifestcheck.internal.util.writeToFile
+import java.io.File
+
+internal class BaselineHandler(private val baselineFile: File) {
+
+    // Only XML baseline files are allowed
+    init {
+        if (!baselineFile.extension.equals("xml", ignoreCase = true)) {
+            throw IllegalArgumentException("The permission baseline has to be a .xml file")
+        }
+    }
+
+    fun serialize(permissions: Map<String, Set<BasePermission>>, recreate: Boolean = false) {
+        val baseline = if (!recreate) deserialize().toMutableMap() else mutableMapOf() // Parse existing baseline and add new or updated permissions
+        permissions.toSortedMap().forEach { (variantName, variantPermissions) ->
+            baseline[variantName] = variantPermissions
+        }
+
+        // Package changes into a XML document
+        val document = createDocumentBuilder()
+            .newDocument()
+        document.appendElement("baseline") {
+            baseline.toSortedMap().forEach { (variantName, variantPermissions) ->
+
+                // Create permission entries for a single variant
+                appendElement("variant") {
+                    setAttribute("name", variantName)
+                    variantPermissions.forEach { permission ->
+                        appendChild(permission.toXmlElement(document))
+                    }
+                }
+            }
+        }
+
+        // Write updated baseline to disk
+        document.writeToFile(baselineFile)
+    }
+
+    fun deserialize(): Map<String, Set<BasePermission>> {
+        val permissions = mutableMapOf<String, Set<BasePermission>>()
+
+        // Missing files are interpreted as empty baselines
+        if (!baselineFile.exists()) {
+            return emptyMap()
+        }
+
+        // Parse the baseline file
+        val document = try {
+            createDocumentBuilder().parse(baselineFile)
+        } catch (exception: Exception) {
+            throw IllegalStateException("Could not parse baseline, please make sure the format is valid", exception)
+        }
+
+        document.getElementsByTagName("variant").forEach { variant ->
+            val variantName = variant.getAttribute("name")
+            val variantPermissions = mutableSetOf<BasePermission>()
+
+            // Parse regular permissions for variant
+            val regularPermissions = variant.getElementsByTagName("uses-permission")
+            regularPermissions.forEach { element ->
+                val name = element.getAttribute("name")
+                val maxSdkVersion = element.getAttribute("maxSdkVersion").toIntOrNull()
+                variantPermissions += Permission(
+                    name,
+                    maxSdkVersion
+                )
+            }
+
+            // Parse SDK 23 permissions for variant
+            val sdk23Permissions = variant.getElementsByTagName("uses-permission-sdk-23")
+            sdk23Permissions.forEach { element ->
+                val name = element.getAttribute("name")
+                val maxSdkVersion = element.getAttribute("maxSdkVersion").toIntOrNull()
+                variantPermissions += Sdk23Permission(
+                    name,
+                    maxSdkVersion
+                )
+            }
+
+            val features = variant.getElementsByTagName("uses-feature")
+            features.forEach { element ->
+                val name = element.getAttribute("name")
+                val required = element.getAttribute("required").toBooleanStrictOrNull()
+                val glEsVersion = element.getAttribute("glEsVersion").takeIf { it.isNotEmpty() }
+                variantPermissions += Feature(
+                    name,
+                    required,
+                    glEsVersion
+                )
+            }
+
+            permissions[variantName] = variantPermissions
+        }
+
+        return permissions
+    }
+}
