+++ plugin-core/src/main/kotlin/com/telefonica/manifestcheck/internal/report/Reporter.kt
@@ -1,0 +1,7 @@
+package com.telefonica.manifestcheck.internal.report
+
+import com.telefonica.manifestcheck.internal.data.Violation
+
+internal interface Reporter {
+    fun report(violations: Map<String, List<Violation>>)
+}
