--- plugin-core/src/main/kotlin/io/github/simonschiller/permissioncheck/internal/report/Reporter.kt
@@ -1,7 +1,0 @@
-package io.github.simonschiller.permissioncheck.internal.report
-
-import io.github.simonschiller.permissioncheck.internal.data.Violation
-
-internal interface Reporter {
-    fun report(violations: Map<String, List<Violation>>)
-}
